requirements.txt
dspy-ai>=2.4.0
datasets>=2.14.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0